Remove commercial specific views support folder.
