+++ commercial/app/views/contentapi/item.scala.html
@@ -1,11 +1,11 @@
-@import _root_.commercial.controllers.{PaidFor, SponsorType, Supported}
+@import commercial.controllers.{PaidFor, SponsorType, Supported}
 @import common.commercial.CardContent
 @import common.LinkTo
 @import fragments.commercial.paidForMeta
 @import views.html.fragments.commercial.cards.itemLargeCard
 @import views.html.fragments.commercial.containerWrapper
 @import views.html.fragments.inlineSvg
-@import views.support.commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
+@import views.support.Commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
 @import views.support.Commercial.glabsLink
 
 @(card: CardContent,
