+++ commercial/app/views/contentapi/items.scala.html
@@ -5,7 +5,7 @@
 @import views.html.fragments.commercial.cards.itemCard
 @import views.html.fragments.commercial.containerWrapper
 @import views.html.fragments.inlineSvg
-@import views.support.commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
+@import views.support.Commercial.TrackingCodeBuilder.mkCapiCardTrackingCode
 @import cards.{Half, Third, Standard}
 @import views.support.Commercial.glabsLink
