+++ common/app/views/fragments/commercial/containers/fixedSmallSlowIII.scala.html
@@ -1,4 +1,4 @@
-@import views.support.commercial.TrackingCodeBuilder.mkInteractionTrackingCode
+@import views.support.Commercial.TrackingCodeBuilder.mkInteractionTrackingCode
 @(frontId: String,
   containerIndex: Int,
   containerModel: common.commercial.ContainerModel)(implicit request: RequestHeader)
